Split up config file into logical parts
